Added View Neighbor for eigrp, Do write to eigrp configuration, and username/password/secret management.
Added a function to ssh into all devices and run a "show ip eigrp neighbors".

Added a "do wr" to the end of the eigrp configuration command list to config is saved whenever routing is setup.

Added  Username, Password, Secret, attributes to the device class and added asking the user for them when scanning or loading an inventory without them. Changed netmiko "deviceInfo" dictionaries to sue the credentials from the device class instances.