Add single device function and improvements for not restconf capable devices
added a function for adding a single device instead of having to scan the entire network. Allows user to input address username password and secret. Still does an ssh test to determine if the device has ssh working and is a cisco type device.

Also fixed some issues where a device not supporting restconf would cause the system to skip some of the ssh based info gathering on those devices.