Full Serial configuration of several devices from factory with serial switch, and tftp utils added
Finished the functionality to use the serial switch to configure device wiht an ip on the appropriate interface to talk to the local tftp and re-apply the stored form the tftp server.

Also added some basic tftp utilities, so copy files to and from a devices flash and the tftp server.